+++ src/app/modules/cooking-recipe/pages/home/home.component.ts
@@ -1,8 +1,17 @@
-import { Component } from '@angular/core';
+import { AfterViewInit, Component, ViewChild } from '@angular/core';
+import { SassHelperComponent } from '../../../../core/utils/components/sass-helper/sass-helper.component';
 
 @Component({
   selector: 'app-home',
   templateUrl: './home.component.html',
   styleUrls: ['./home.component.sass'],
 })
-export class HomeComponent {}
+export class HomeComponent implements AfterViewInit {
+  @ViewChild(SassHelperComponent) appSassHelper!: SassHelperComponent;
+
+  ngAfterViewInit() {
+    if (this.appSassHelper) {
+      console.log(this.appSassHelper.readProperty('primary--background'));
+    }
+  }
+}